+++ altrpnjs/helpers/delete.ts
@@ -1,6 +1,7 @@
 import { DateTime } from 'luxon';
-import {BaseModel, ModelQueryBuilderContract} from "@ioc:Adonis/Lucid/Orm";
+import {BaseModel, ModelQueryBuilderContract, } from "@ioc:Adonis/Lucid/Orm";
 import * as _ from 'lodash'
+import Database from "@ioc:Adonis/Lucid/Database";
 
 export const softDeleteQuery = (query: ModelQueryBuilderContract<typeof BaseModel> | [ModelQueryBuilderContract<typeof BaseModel>, ModelQueryBuilderContract<typeof BaseModel>]) => {
   if(_.isArray(query)){
@@ -11,12 +12,17 @@ export const softDeleteQuery = (query: ModelQueryBuilderContract<typeof BaseMode
   }
 }
 export const softDelete = async (instance) => {
-  // @ts-ignore
-  instance.deleted_at = DateTime.local();
-  // @ts-ignore
-  instance.deletedAt = DateTime.local();
-  // @ts-ignore
-  await instance.save();
+
+  const Model = instance.constructor;
+  await Model.$hooks.exec('before', 'delete', instance);
+
+  await  Database
+    .from(Model.table) // Название таблицы из переменной
+    .where('id', instance.id) // Поиск по колонке id
+    .update({ deleted_at: DateTime.local() }) // Обновление значения deleted_at
+
+  await Model.$hooks.exec('after', 'delete', instance);
+
 }
 
 export const beforePaginateQuery = (