+++ resources/modules/admin/src/components/TopMemoryUsage.js
@@ -16,7 +16,12 @@ class TopMemoryUsage extends Component{
       if(testEnable){
         this.setState(state=>({...state,
           updating:true,}))
-        const res = await this.resource.getAll()
+        try {
+          const res = await this.resource.getAll()
+
+        }catch (e) {
+          console.error(e)
+        }
         this.setState(state=>({...state,
           updating:false,
           used: res.data.used,}))